# performance: optimize the performance of interaction.