--- a/src/animation/group-action.js
+++ b/src/animation/group-action.js
@@ -3,7 +3,6 @@
  * @author sima.zhang
  */
 const Util = require('./util');
-const Helper = require('../util/helper');
 const { Shape } = require('../graphic/index');
 
 function _groupScaleIn(container, animateCfg, coord, zeroY, type) {
@@ -92,34 +91,14 @@ function shapesScaleInXY(container, animateCfg) {
 }
 
 function groupWaveIn(container, animateCfg, coord) {
-  const clip = Helper.getClip(coord);
+  const clip = Util.getClip(coord);
   clip.set('canvas', container.get('canvas'));
   container.attr('clip', clip);
   const onEnd = function() {
     container.attr('clip', null);
     clip.remove(true);
   };
-
-  const endState = {};
-  if (coord.isPolar) {
-    const { startAngle, endAngle } = coord;
-    endState.endAngle = endAngle;
-    clip.attr('endAngle', startAngle);
-  } else {
-    const { start, end } = coord;
-    const width = Math.abs(start.x - end.x);
-    const height = Math.abs(start.y - end.y);
-
-    if (coord.isTransposed) {
-      clip.attr('height', 0);
-      endState.height = height;
-    } else {
-      clip.attr('width', 0);
-      endState.width = width;
-    }
-  }
-
-  Util.doAnimation(clip, endState, animateCfg, onEnd);
+  Util.doAnimation(clip, clip.endState, animateCfg, onEnd);
 }
 
 module.exports = {