<!--
Thank you for your pull request. Please review below requirements.
Bug fixes and new features should include tests and possibly benchmarks.
Contributors guide: https://github.com/antvis/F2/blob/master/CONTRIBUTING.zh-CN.md

感谢您贡献代码。请确认下列 checklist 的完成情况。
Bug 修复和新功能必须包含测试，必要时请附上性能测试。
Contributors guide: https://github.com/antvis/f2/blob/master/CONTRIBUTING.zh-CN.md
-->

##### Checklist
<!-- Remove items that do not apply. For completed items, change [ ] to [x]. -->

- [ ] `npm test` passes
- [ ] tests and/or benchmarks are included
- [ ] commit message follows commit guidelines

##### Description of change
<!-- Provide a description of the change below this comment. -->

pan 和 pinch 的交互原理是在全量数据载入的基础上，通过操作相关的  scale 达成的，所以就会造成一些不在 scale 范围内的数据参与了图形的创建、绘制以及 tooltip 等的查找过程，造成了一定的性能消耗，此次优化主要针对这个场景，以下优化前后的数据（通过线图）

---

## 场景一

![image](https://user-images.githubusercontent.com/6628666/43565780-94470e50-965e-11e8-8a0f-0ddd760b2b63.png)

* 图表类型：折线图
* 数据条数： 1164
* 首次绘制数据条数： 1164

评估此次优化对正常图表（不带交互）绘制的影响，看前后的渲染时间并没有多大差异。

<div class="bi-table">
  <table>
    <colgroup>
      <col width="auto" />
      <col width="auto" />
      <col width="auto" />
    </colgroup>
    <tbody>
      <tr height="34px">
        <td rowspan="1" colSpan="1">
          <div data-type="alignment" data-value="center" style="text-align:center">
            <div data-type="p"><strong>场景</strong></div>
          </div>
        </td>
        <td rowspan="1" colSpan="1">
          <div data-type="alignment" data-value="center" style="text-align:center">
            <div data-type="p"><strong>优化前</strong></div>
          </div>
        </td>
        <td rowspan="1" colSpan="1">
          <div data-type="alignment" data-value="center" style="text-align:center">
            <div data-type="p"><strong>优化后</strong></div>
          </div>
        </td>
      </tr>
      <tr height="34px">
        <td rowspan="1" colSpan="1">
          <div data-type="p"><span data-type="color" style="color:#FA541C">render</span></div>
        </td>
        <td rowspan="1" colSpan="1">
          <div data-type="p">33.6 ms</div>
        </td>
        <td rowspan="1" colSpan="1">
          <div data-type="p">34.1 ms</div>
        </td>
      </tr>
    </tbody>
  </table>
</div>


---

## 场景二

![image](https://user-images.githubusercontent.com/6628666/43565800-a64b355e-965e-11e8-9d50-2a4cdd3575bf.png)

* 图表类型： 柱状图
* 数据条数：200
* 首次绘制条数： 21
* x 轴数据类型： 'timeCat'

以下是渲染时间对比：

<div class="bi-table">
  <table>
    <colgroup>
      <col width="auto" />
      <col width="auto" />
      <col width="auto" />
    </colgroup>
    <tbody>
      <tr height="34px">
        <td rowspan="1" colSpan="1">
          <div data-type="alignment" data-value="center" style="text-align:center">
            <div data-type="p"><strong>场景</strong></div>
          </div>
        </td>
        <td rowspan="1" colSpan="1">
          <div data-type="alignment" data-value="center" style="text-align:center">
            <div data-type="p"><strong>优化前</strong></div>
          </div>
        </td>
        <td rowspan="1" colSpan="1">
          <div data-type="alignment" data-value="center" style="text-align:center">
            <div data-type="p"><strong>优化后</strong></div>
          </div>
        </td>
      </tr>
      <tr height="34px">
        <td rowspan="1" colSpan="1">
          <div data-type="p"><span data-type="color" style="color:#FA541C">render</span></div>
        </td>
        <td rowspan="1" colSpan="1">
          <div data-type="p"> 28.3 ms</div>
        </td>
        <td rowspan="1" colSpan="1">
          <div data-type="p">22.5 ms</div>
        </td>
      </tr>
      <tr height="34px">
        <td rowspan="1" colSpan="1">
          <div data-type="p"><span data-type="color" style="color:#FA541C">tooltip 查找</span></div>
        </td>
        <td rowspan="1" colSpan="1">
          <div data-type="p"> 8.7 ms</div>
        </td>
        <td rowspan="1" colSpan="1">
          <div data-type="p">3.0 ms</div>
        </td>
      </tr>
      <tr height="34px">
        <td rowspan="1" colSpan="1">
          <div data-type="p"><span data-type="color" style="color:#FA541C">平移时每次重绘的时间</span></div>
        </td>
        <td rowspan="1" colSpan="1">
          <div data-type="p">5 ~ 9 ms</div>
        </td>
        <td rowspan="1" colSpan="1">
          <div data-type="p">3 ~ 5 m</div>
        </td>
      </tr>
    </tbody>
  </table>
</div>


---

## 场景三

![image](https://user-images.githubusercontent.com/6628666/43565804-ae1b27da-965e-11e8-9572-56160880cd10.png)


* 图表类型：折线图
* 数据条数： 1164
* 首次绘制数据条数： 120
* x 轴数据类型： 'cat'

| __场景__ | __优化前__ | __优化后__ |
| :--- | :--- | :--- |
| <span data-type="color" style="color:rgb(250, 84, 28)">render</span> | 31.5 ms | 25 ms |
| <span data-type="color" style="color:rgb(250, 84, 28)">tooltip 查找</span> | 3.5 ~ 4.2 ms | 2.8 ~ 3.1 ms |
| <span data-type="color" style="color:rgb(250, 84, 28)">平移时每次重绘的时间</span> | 8 ~ 12 ms | 5 ~ 8 ms |